clean up enrollment contexts.
This is a mess.

For now: Enroll, & Drop are children of Catalog list, so always drill up to it. The links on the detail page are just helpfull. Gifting is a child of the detail view. drill up to it.

Long game: Catalog and Enrollment are one in the same. merge packages together. Fix routers so contexts work for us instead of against us.