cleanup lint. As of React 0.14, you can define stateless components which are simply a function with a single argument (props object) and return a "render" -- a JSX or React.createElement expression.
eslint-react-plugin doesn't try to guess if a function is a stateless component or a helper... and validates all functions that return a "render" as a stateless component. This is normally fine, but the display-name rule by default wants a "static" property "displayName"... but allows through configuration to trust the function name as the displayName. (a la "Function.name") ... we've turned this one, the one caveat is that we cannot have "anonymous" functions that return "renders"